Move emailsReponse declaration to Api
